Coding - Scope 20 AIS enums as enum class
Converts 20 unscoped AIS enums to enum class:
AIS_TypeOfPlane, AIS_TypeOfAttribute, AIS_NavigationMode,
AIS_StatusOfPick, AIS_SelectionScheme, AIS_SelectionModesConcurrency,
AIS_SelectStatus, AIS_KindOfInteractive, AIS_TrihedronSelectionMode,
AIS_ManipulatorMode, AIS_TypeOfAxis, AIS_WalkTranslation, AIS_WalkRotation,
AIS_TypeOfIso, AIS_MouseGesture, AIS_RotationMode, AIS_ViewSelectionTool,
AIS_ViewInputBufferType, AIS_StatusOfDetection, AIS_DragAction.

AIS_DisplayMode is intentionally kept as a plain enum because its
values are passed to `int` display-mode parameters throughout the
viewer API; converting it to enum class would require static_cast
at ~40 call sites for no clean type-safety win.

Localized casts added where enum values interact with int APIs:
SelectionPriority storage, NCollection_DataMap<int, ...> keys for
AIS_KindOfInteractive, AIS_Manipulator context methods taking int
modes, AIS_Trihedron selection-mode switch cases, and array
subscripts in AIS_WalkDelta.

+++ src/Visualization/TKV3d/AIS/AIS_ExclusionFilter.cxx
@@ -124,7 +124,7 @@ bool AIS_ExclusionFilter::IsStored(const AIS_KindOfInteractive aType) const
 bool AIS_ExclusionFilter::IsSignatureIn(const AIS_KindOfInteractive aType,
                                         const int                   SignatureInType) const
 {
-  if (!myStoredTypes.IsBound(aType))
+  if (!myStoredTypes.IsBound(static_cast<int>(aType)))
     return false;
   for (NCollection_List<int>::Iterator Lit(myStoredTypes((int)aType)); Lit.More(); Lit.Next())
   {
@@ -151,7 +151,8 @@ void AIS_ExclusionFilter::ListOfSignature(const AIS_KindOfInteractive aType,
 {
   TheStoredList.Clear();
   if (IsStored(aType))
-    for (NCollection_List<int>::Iterator it(myStoredTypes(aType)); it.More(); it.Next())
+    for (NCollection_List<int>::Iterator it(myStoredTypes(static_cast<int>(aType))); it.More();
+         it.Next())
       TheStoredList.Append(it.Value());
 }
 
@@ -169,10 +170,10 @@ bool AIS_ExclusionFilter::IsOk(const occ::handle<SelectMgr_EntityOwner>& EO) con
     return false;
 
   // type of AIS is not in the map...
-  if (!myStoredTypes.IsBound(IO->Type()))
+  if (!myStoredTypes.IsBound(static_cast<int>(IO->Type())))
     return myIsExclusionFlagOn;
   // type of AIS is not in the map and there is no signature indicated
-  if (myStoredTypes(IO->Type()).IsEmpty())
+  if (myStoredTypes(static_cast<int>(IO->Type())).IsEmpty())
     return !myIsExclusionFlagOn;
   // one or several signatures are indicated...
   if (IsSignatureIn(IO->Type(), IO->Signature()))
